subir notas

+++ src/app/backend/admin/tableCourses.php
@@ -15,8 +15,8 @@
         id INT AUTO_INCREMENT PRIMARY KEY,
         nombre VARCHAR(32),
         codigo INT(7),
-        nota1 FLOAT,
-        notaFinal FLOAT
+        nota1 decimal(10, 1),
+        notaFinal decimal(10, 2)
     )";
 
     $resultado = mysqli_query($conexion, $consulta);
